Switch pages so financial accounts are 3 pages and only the page that needs context uses it to prevent rerenders on the other pages.
